Move catalog config endpoint contributor to core
