+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/generic/GenericTableConfigEndpoints.java
@@ -25,14 +25,14 @@ import java.util.Set;
 import org.apache.iceberg.rest.Endpoint;
 import org.apache.polaris.core.config.FeatureConfiguration;
 import org.apache.polaris.core.config.RealmConfig;
+import org.apache.polaris.core.rest.CatalogConfigEndpointContributor;
 import org.apache.polaris.core.rest.GenericTableEndpoints;
-import org.apache.polaris.service.catalog.config.CatalogConfigEndpointContributor;
 
 @ApplicationScoped
 public class GenericTableConfigEndpoints {
   @Produces
-  public CatalogConfigEndpointContributor genericTableEndpoints() {
-    return GenericTableConfigEndpoints::getSupportedGenericTableEndpoints;
+  public CatalogConfigEndpointContributor genericTableEndpoints(RealmConfig realmConfig) {
+    return () -> getSupportedGenericTableEndpoints(realmConfig);
   }
 
   /**
