+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/policy/PolicyConfigEndpoints.java
@@ -25,13 +25,13 @@ import java.util.Set;
 import org.apache.iceberg.rest.Endpoint;
 import org.apache.polaris.core.config.FeatureConfiguration;
 import org.apache.polaris.core.config.RealmConfig;
-import org.apache.polaris.service.catalog.config.CatalogConfigEndpointContributor;
+import org.apache.polaris.core.rest.CatalogConfigEndpointContributor;
 
 @ApplicationScoped
 public class PolicyConfigEndpoints {
   @Produces
-  public CatalogConfigEndpointContributor policyEndpoints() {
-    return PolicyConfigEndpoints::getSupportedPolicyEndpoints;
+  public CatalogConfigEndpointContributor policyEndpoints(RealmConfig realmConfig) {
+    return () -> getSupportedPolicyEndpoints(realmConfig);
   }
 
   /**
